requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
selenium~=4.27.1
nltk~=3.9.1
pydub~=0.25.1
openai-whisper
